properly rename to FeatureCollection
