+++ src/IIIF/IIIF.Tests/Serialisation/ManifestSerialisationTests.cs
@@ -7,7 +7,7 @@ using IIIF.ImageApi.V3;
 using IIIF.Presentation.V3;
 using IIIF.Presentation.V3.Annotation;
 using IIIF.Presentation.V3.Content;
-using IIIF.Presentation.V3.FeatureCollection;
+using IIIF.Presentation.V3.NavPlace;
 using IIIF.Presentation.V3.Strings;
 using IIIF.Serialisation;
 using Newtonsoft.Json.Linq;
@@ -200,7 +200,7 @@ public class ManifestSerialisationTests
                     Format = "application/json"
                 }
             },
-            NavPlace = new NavPlace()
+            NavPlace = new FeatureCollection()
             {
                 Id = "https://test.example.com/nav-place",
                 Features = new List<Feature>()
