+++ src/IIIF/IIIF/Presentation/V3/NavPlace/Feature.cs
@@ -1,4 +1,4 @@
-namespace IIIF.Presentation.V3.FeatureCollection;
+namespace IIIF.Presentation.V3.NavPlace;
 
 public class Feature : IResource
 {
