+++ src/IIIF/IIIF/Presentation/V3/NavPlace/Geometry.cs
@@ -1,4 +1,4 @@
-namespace IIIF.Presentation.V3.FeatureCollection;
+namespace IIIF.Presentation.V3.NavPlace;
 
 /// <summary>
 /// Base class for a Geometry
