+++ src/IIIF/IIIF/Presentation/V3/NavPlace/MultiPoint.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace IIIF.Presentation.V3.FeatureCollection;
+namespace IIIF.Presentation.V3.NavPlace;
 
 public class MultiPoint : Geometry
 {
